split code into modules for unixes and windows
