+++ src/windows/mod.rs
@@ -1,0 +1,7 @@
+pub fn find_af_inet<'a>() -> Result<Vec<(&'a str, IpAddr)>, Error> {
+    //TODO: implement this on windows
+}
+
+pub fn local_ip() -> Result<IpAddr, Error> {
+    //TODO: implement this on windows
+}
